+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 - Implement Giphy actions handler
 - Fix .gif preview rendering on message list 
 - Fix thread shown issue after sending message to a channel 
-
 - Remove border related attributes from MessageInputView. Add close button background attribute to MessageInputView.
+- Improve setting user in sample app
 
 # Sep 4th, 2020 - 4.2.11-beta-11